Migrate to CMake
The built-in GoogleTest  Adapter for Visual Studio is an older version that no longer supports Gmock after google separated the libraries. Previous build system included a nasty workaround and many manual steps to setup and was a pain.  The only full solution to provide true cross-platform compatibility and eliminate issues with the build was to migrate to a different system entirely.